requirements.txt
azure-ai-formrecognizer==3.1.2